requirements.txt
podsearch==0.3.1
pyradios==1.0.2
python_dateutil==2.8.2
pytube==12.1.0
rich==12.5.1
scipy==1.8.1
ShazamAPI==0.0.2
sounddevice==0.4.4
spotdl==4.0.0-rc.2
SQLAlchemy==1.4.37
yt_dlp==2022.6.29
